[FEAT] Add splash page
